refactor: always enable rayon
