+++ Cargo.toml
@@ -16,8 +16,8 @@ version = "0.1.0"
 cfg-if = "1.0"
 num-traits = { version = "0.2", default-features = false }
 nonzero_ext = "0.3"
-im = "15.1.0"
-rayon = { version = "1.10", optional = true }
+im = { version = "15.1.0", features = ["rayon"] }
+rayon = "1.10"
 
 [dev-dependencies]
 quickcheck = "1.0"
@@ -29,9 +29,8 @@ path = "benches/benches.rs"
 harness = false
 
 [features]
-default = ["std", "rayon"]
+default = ["std"]
 std = ["num-traits/std"]
-rayon = ["im/rayon", "dep:rayon"]
 
 [profile.bench]
 debug = true
